Release prep: Moved routing strategies to their own namespace
