+++ examples/helloworld/api.php
@@ -31,7 +31,7 @@ use gordonmcvey\WarpCore\FrontController;
 use gordonmcvey\WarpCore\middleware\CallStackFactory;
 use gordonmcvey\WarpCore\routing\RequestPathValidator;
 use gordonmcvey\WarpCore\routing\Router;
-use gordonmcvey\WarpCore\routing\SingleControllerStrategy;
+use gordonmcvey\WarpCore\routing\strategy\SingleControllerStrategy;
 use gordonmcvey\WarpCore\sdk\interface\controller\RequestHandlerInterface;
 
 /**
