+++ examples/psr7/api.php
@@ -1,5 +1,7 @@
 <?php
 
+/** @noinspection ALL */
+
 /**
  * Copyright © 2015 Docnet, 2025 Gordon McVey
  *
@@ -30,7 +32,7 @@ use gordonmcvey\WarpCore\FrontController;
 use gordonmcvey\WarpCore\middleware\CallStackFactory;
 use gordonmcvey\WarpCore\routing\RequestPathValidator;
 use gordonmcvey\WarpCore\routing\Router;
-use gordonmcvey\WarpCore\routing\SingleControllerStrategy;
+use gordonmcvey\WarpCore\routing\strategy\SingleControllerStrategy;
 use GuzzleHttp\Psr7\ServerRequest;
 use GuzzleHttp\Psr7\Utils;
