+++ src/routing/strategy/PathNamespaceStrategy.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace gordonmcvey\WarpCore\routing;
+namespace gordonmcvey\WarpCore\routing\strategy;
 
 use gordonmcvey\httpsupport\enum\Verbs;
 use gordonmcvey\WarpCore\sdk\interface\routing\RoutingStrategyInterface;
