+++ tests/integration/FrontControllerTest.php
@@ -32,10 +32,10 @@ use gordonmcvey\WarpCore\controller\ControllerFactory;
 use gordonmcvey\WarpCore\error\JsonErrorHandler;
 use gordonmcvey\WarpCore\FrontController;
 use gordonmcvey\WarpCore\middleware\CallStackFactory;
-use gordonmcvey\WarpCore\routing\PathNamespaceStrategy;
 use gordonmcvey\WarpCore\routing\RequestPathValidator;
 use gordonmcvey\WarpCore\routing\Router;
-use gordonmcvey\WarpCore\routing\SingleControllerStrategy;
+use gordonmcvey\WarpCore\routing\strategy\PathNamespaceStrategy;
+use gordonmcvey\WarpCore\routing\strategy\SingleControllerStrategy;
 use gordonmcvey\WarpCore\sdk\exception\AccessDenied;
 use gordonmcvey\WarpCore\sdk\exception\Auth;
 use gordonmcvey\WarpCore\sdk\exception\controller\BootstrapFailure;
