+++ tests/integration/routing/RouterTest.php
@@ -7,10 +7,10 @@ namespace gordonmcvey\WarpCore\test\integration\routing;
 use gordonmcvey\httpsupport\enum\statuscodes\ClientErrorCodes;
 use gordonmcvey\httpsupport\enum\Verbs;
 use gordonmcvey\httpsupport\interface\request\RequestInterface;
-use gordonmcvey\WarpCore\routing\PathNamespaceStrategy;
 use gordonmcvey\WarpCore\routing\RequestPathValidator;
 use gordonmcvey\WarpCore\routing\Router;
-use gordonmcvey\WarpCore\routing\StaticStrategy;
+use gordonmcvey\WarpCore\routing\strategy\PathNamespaceStrategy;
+use gordonmcvey\WarpCore\routing\strategy\StaticStrategy;
 use gordonmcvey\WarpCore\sdk\exception\routing\MethodNotAllowed;
 use gordonmcvey\WarpCore\sdk\exception\routing\NoRouteToController;
 use PHPUnit\Framework\Attributes\DataProvider;
