+++ tests/unit/routing/strategy/SingleControllerStrategyTest.php
@@ -18,10 +18,10 @@
 
 declare(strict_types=1);
 
-namespace gordonmcvey\WarpCore\test\unit\routing;
+namespace gordonmcvey\WarpCore\test\unit\routing\strategy;
 
 use gordonmcvey\httpsupport\enum\Verbs;
-use gordonmcvey\WarpCore\routing\SingleControllerStrategy;
+use gordonmcvey\WarpCore\routing\strategy\SingleControllerStrategy;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
